# feat: Implement comprehensive quality gates and security framework