--- a/.bandit
+++ b/.bandit
@@ -1,0 +1,36 @@
+[bandit]
+# Bandit security configuration for LeanVibe Quality Agent
+
+# Directories to exclude from scanning
+exclude_dirs = [
+    ".git",
+    ".pytest_cache",
+    "__pycache__",
+    "htmlcov",
+    "test_reports",
+    "tutorials"
+]
+
+# Test IDs to skip (if needed)
+skips = [
+    # B101: assert_used - Allow asserts in tests
+    "B101"
+]
+
+# Test severities (LOW, MEDIUM, HIGH)
+# Only report MEDIUM and HIGH severity issues
+severity = MEDIUM
+
+# Confidence levels (LOW, MEDIUM, HIGH)
+# Only report HIGH confidence issues to reduce false positives
+confidence = HIGH
+
+# Output format
+format = json
+output = analysis_reports/bandit_security_report.json
+
+# Include test files for security analysis
+tests = ["tests"]
+
+# Recursive scanning
+recursive = true