--- a/mypy.ini
+++ b/mypy.ini
@@ -1,0 +1,52 @@
+[mypy]
+# MyPy configuration for LeanVibe Quality Agent
+
+# Type checking strictness
+python_version = 3.13
+warn_return_any = True
+warn_unused_configs = True
+disallow_untyped_defs = False
+disallow_incomplete_defs = True
+check_untyped_defs = True
+disallow_untyped_decorators = False
+
+# Import discovery
+mypy_path = .claude
+namespace_packages = True
+explicit_package_bases = True
+
+# Error output
+show_error_codes = True
+show_column_numbers = True
+show_error_context = True
+color_output = True
+
+# Warnings
+warn_redundant_casts = True
+warn_unused_ignores = True
+warn_no_return = True
+warn_unreachable = True
+
+# Module-specific configurations
+[mypy-.claude.*]
+ignore_missing_imports = True
+
+[mypy-tests.*]
+ignore_missing_imports = True
+disallow_untyped_defs = False
+
+# Third-party modules
+[mypy-pandas.*]
+ignore_missing_imports = True
+
+[mypy-numpy.*]
+ignore_missing_imports = True
+
+[mypy-sklearn.*]
+ignore_missing_imports = True
+
+[mypy-psutil.*]
+ignore_missing_imports = True
+
+[mypy-yaml.*]
+ignore_missing_imports = True