--- a/requirements.txt
+++ b/requirements.txt
@@ -13,6 +13,11 @@ pandas==2.0.3
 # Development and testing
 black>=23.0
 ruff>=0.0.290
+pytest-cov>=4.0.0
+coverage>=7.0.0
+mypy>=1.0.0
+bandit>=1.7.5
+safety>=2.0.0
 
 # Configuration and utilities
 python-dotenv>=1.0.0