--- a/tests/mutation/__init__.py
+++ b/tests/mutation/__init__.py
@@ -1,0 +1,1 @@
+"""Mutation testing module for LeanVibe Quality Agent."""