--- a/tests/performance/__init__.py
+++ b/tests/performance/__init__.py
@@ -1,1 +1,1 @@
-"""Performance tests package."""
+"""Performance testing module for LeanVibe Quality Agent."""