--- a/tests/reporting/__init__.py
+++ b/tests/reporting/__init__.py
@@ -1,0 +1,1 @@
+"""Automated test reporting module for LeanVibe Quality Agent."""