… .github/workflows/dependency-security-monitoring.yml
… .pre-commit-config.yaml
… CLAUDE.md
… QUALITY_AGENT_IMPLEMENTATION_SUMMARY.md
… QUALITY_FRAMEWORK_SUMMARY.md
… agent_CLAUDE.md
… analysis_reports/bandit_security_report.json
… analysis_reports/safety_report.json
… config/security_monitoring.json
… quality_monitoring.log
… quality_reports/quality_analysis_20250715_185830.json
… quality_reports/quality_dashboard_20250715_185830.html
… scripts/quality_metrics_monitor.py
… scripts/security_monitoring.py
… security_monitoring.log
… security_reports/security_scan_20250715_185432.json
… tests/distributed/test_distributed_framework.py
… tests/integration/test_system_integration.py
… tests/mutation/test_mutation_framework.py
… tests/performance/test_performance_benchmarks.py
… tests/reporting/test_report_generator.py
… tests/security/test_security_framework.py